refactor(client): inputページ部品コンポーネント化
