+++ client/src/app/_component/index.ts
@@ -1,3 +1,2 @@
 export * from "./modal";
 export * from "./morphoto-header";
-export * from "./slider";
