+++ client/src/app/input/_components/file-uploader/index.ts
@@ -1,0 +1,1 @@
+export * from "./file-uploader";
