+++ client/src/app/input/_components/index.ts
@@ -1,3 +1,6 @@
+export * from "./file-uploader";
 export * from "./input-button";
 export * from "./prompt-card";
+export * from "./slider";
 export * from "./title";
+export * from "./upload-card";
