+++ client/src/app/input/_components/slider/index.ts
