+++ client/src/app/input/_components/slider/slider.tsx
