+++ client/src/app/input/_components/upload-card/index.ts
@@ -1,0 +1,1 @@
+export * from "./upload-card";
